requirements.txt
fastapi==0.116.1
httpx==0.28.1
openai==1.99.9
opencv_python_headless==4.11.0.86
pydantic==2.11.7
pydantic_settings==2.10.1
PyYAML==6.0.2
PyYAML==6.0.2
SQLAlchemy==2.0.28
uvicorn==0.35.0
